+++ apps/client/src/app/components/user-detail-dialog/interfaces/interfaces.ts
@@ -2,6 +2,5 @@ import { AdminUsers } from '@ghostfolio/common/interfaces';
 
 export interface UserDetailDialogParams {
   deviceType: string;
-  userData?: AdminUsers['users'][0];
-  userId: string;
+  userData: AdminUsers['users'][0];
 }