+++ CHANGELOG.md
@@ -5,12 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+- User detail dialog component to view user information in admin users management
+
 ## 2.210.0 - 2025-10-22
 
 ### Added
 
 - Added support for data gathering by date range in the asset profile details dialog of the admin control panel
-- Added user detail dialog component to view user information in admin users management
 
 ### Changed